updated Jul 2023 to remove identifying information
I haven't tested these files recently.  This was before I had any formal computer science training, so it's likely that there's a good bit of hardcoding that will cause it to break in environments other than the one it was written for.  Additionally, the removal of identifying information was done somewhat quickly and could potentially have introduced bugs.